requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
coinpaprika-sdk==1.0.0
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
feedparser==6.0.11
fonttools==4.54.1
frozenlist==1.4.1
fuzzywuzzy==0.18.0
greenlet==3.2.3
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_resources==6.4.5
iso8601==2.1.0
jiter==0.6.1
kiwisolver==1.4.7
Levenshtein==0.26.0
matplotlib==3.9.2
multidict==6.1.0
numpy==2.0.2
openai==0.28.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==11.0.0
propcache==0.2.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.26.0
pycoingecko==3.1.0
pytz==2024.2
RapidFuzz==3.10.0
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tabulate==0.9.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.15
vaderSentiment==3.3.2
yarl==1.15.5
zipp==3.20.2
